refactor(app-shell): deepen renderer application orchestration
